Add rudimentary support for tags on collections
This is a little bit hacky, because Jekyll doesn't actually have tags on
collections, but we can fake it for the purposes of pagination.

Usage:
1. Add a property called 'tags' to the documents in the collection.

2. Use the 'collection_tag' property in paginator to filter on one tag
from the list.

Note that it just does string matching for now, rather than actually
converting the tags on the collection documents to an array or anything
sensible like that.