Fixes deprecation warning on categories and tags
At some point, Jekyll moved the categories and tags for a post into the
data hash, deprecating the accessors. Using p.categories or p.tags
for accessing the list of categories or tags for a post p has been
deprecated for a while, being the data hash p['categories'] / p['posts']
the preferred way for accesing those parameters.

This commit uses the data hash instead of the accessors to avoid the
noisy deprecation warnings that log a lot of warnings during the
compilation of a site.